+++ typings/index.d.ts
@@ -5,6 +5,8 @@ import { Object as ObjectTB } from 'ts-toolbelt';
 export type Value = null | string | number | object;
 export type Extension = { [key in PropertyKey] : unknown };
 
+export declare const proxyKey : unique symbol;
+
 export type Proxied<V extends Value, E extends Extension = {}> =
     ObjectTB.Merge<
         Readonly<ObjectTB.Merge<
@@ -20,10 +22,16 @@ export type Proxied<V extends Value, E extends Extension = {}> =
             : V
     >;
 
-export declare const proxyKey : unique symbol;
+export declare const unwrap : <V extends Value, E extends Extension>(proxy : Proxied<V, E>) => {
+    value : V,
+    extension : E,
+};
 
-export declare const extend : <V extends Value, E extends Extension = {}>(value : V, extension ?: E)
-    => Proxied<V, E>;
+export declare const extend :
+    (<V extends Value, E extends Extension = {}>(value : V, extension ?: E) => Proxied<V, E>) & {
+        proxyKey : typeof proxyKey,
+        unwrap : typeof unwrap,
+    };
 
 export declare const registerProxyFormatter : () => void;